+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 
    * based on MeshLab 202y.xx
    * all filters renamed, old filter names are now deprecated
+   * throw exception when a parameter name is not found (was just a print before)
    * bugfixes
 
 2021.10